HAL_ChibiOS: port SPI, I2C and UART drivers to H7

+++ libraries/AP_HAL_ChibiOS/SPIDevice.h
@@ -48,7 +48,7 @@ public:
 struct SPIDesc {
     SPIDesc(const char *_name, uint8_t _bus,
             uint8_t _device, ioline_t _pal_line,
-            uint16_t _mode, uint32_t _lowspeed, uint32_t _highspeed)
+            uint32_t _mode, uint32_t _lowspeed, uint32_t _highspeed)
         : name(_name), bus(_bus), device(_device),
           pal_line(_pal_line), mode(_mode),
           lowspeed(_lowspeed), highspeed(_highspeed)
@@ -59,7 +59,7 @@ struct SPIDesc {
     uint8_t bus;
     uint8_t device;
     ioline_t pal_line;
-    uint16_t mode;
+    uint32_t mode;
     uint32_t lowspeed;
     uint32_t highspeed;
 };
@@ -119,14 +119,14 @@ private:
     SPIBus &bus;
     SPIDesc &device_desc;
     uint32_t frequency;
-    uint16_t freq_flag;
-    uint16_t freq_flag_low;
-    uint16_t freq_flag_high;
+    uint32_t freq_flag;
+    uint32_t freq_flag_low;
+    uint32_t freq_flag_high;
     char *pname;
     bool cs_forced;
     static void *spi_thread(void *arg);
-    static uint16_t derive_freq_flag_bus(uint8_t busid, uint32_t _frequency);
-    uint16_t derive_freq_flag(uint32_t _frequency);
+    static uint32_t derive_freq_flag_bus(uint8_t busid, uint32_t _frequency);
+    uint32_t derive_freq_flag(uint32_t _frequency);
 };
 
 class SPIDeviceManager : public AP_HAL::SPIDeviceManager {
